requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
gigachat
python-dotenv